moved run logic to runner package - moved finder logic to loader - moved parser call from main to loader
